+++ compiler/Acton/CommandLineParser.hs
@@ -60,8 +60,8 @@ data CompileOptions   = CompileOptions {
                          quiet       :: Bool,
                          debug       :: Bool,
                          dev         :: Bool,
+                         only_build  :: Bool,
                          skip_build  :: Bool,
---                         only_c      :: Bool,
                          root        :: String,
                          tempdir     :: String,
                          syspath     :: String,
@@ -77,6 +77,7 @@ data BuildOptions = BuildOptions {
                          dbB         :: Bool,
                          debugB      :: Bool,
                          devB        :: Bool,
+                         only_buildB :: Bool,
                          skip_buildB :: Bool,
                          autostubB   :: Bool,
                          rootB       :: String,
@@ -153,8 +154,8 @@ compileOptions = CompileOptions
         <*> switch (long "quiet"        <> help "Don't print stuff")
         <*> switch (long "debug"        <> help "Print debug stuff")
         <*> switch (long "dev"          <> help "Development mode; include debug symbols etc")
+        <*> switch (long "only-build"   <> help "Only perform final build of .c files, do not compile .act files")
         <*> switch (long "skip-build"   <> help "Skip final bulid of .c files")
---        <*> switch (long "only-c"       <> help "Just compile .c files, skip .act compilation (might mean using outdated .c files)")
         <*> strOption (long "root"      <> metavar "ROOTACTOR" <> value "" <> help "Set root actor")
         <*> strOption (long "tempdir"   <> metavar "TEMPDIR" <> value "" <> help "Set directory for build files")
         <*> strOption (long "syspath"   <> metavar "TARGETDIR" <> value "" <> help "Set syspath")
@@ -170,6 +171,7 @@ buildCommand          = Build <$> (
         <*> switch (long "db"           <> help "Enable DB backend")
         <*> switch (long "debug"        <> help "Print debug stuff")
         <*> switch (long "dev"          <> help "Development mode; include debug symbols etc")
+        <*> switch (long "only-build"   <> help "Only perform final build of .c files, do not compile .act files")
         <*> switch (long "skip-build"   <> help "Skip final build of .c files")
         <*> switch (long "auto-stub"    <> help "Allow automatic stub detection")
         <*> strOption (long "root"      <> metavar "ROOTACTOR" <> value "" <> help "Set root actor")